+++ NEWS.md
@@ -9,6 +9,7 @@
 * `step_unorder` converts ordered factors to unordered factors. 
 * `step_count` counts the number of instances that a pattern exists in a string. 
 * `step_string2factor` and `step_factor2string` can be used to move between encodings. 
+* `step_lowerimpute` is for numeric data where the values cannot be measured below a specific value. For these cases, random uniform values are used for the truncated values.  
 * A step to remove simple zero-variance variables was added (`step_zv`).
 * A series of `tidy` methods were added for recipes and many (but not all) steps.